Removed newly created BorderThicknessToStrokeThicknessConverter and moved the written code to existing ThicknessToDoubleConverter. Removed ThicknessToDoubleConverter resource entries where the static resource is not referenced.
