+++ src/MahApps.Metro/MahApps.Metro/Styles/Controls.PasswordBox.xaml
@@ -10,7 +10,6 @@
     </ResourceDictionary.MergedDictionaries>
 
     <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
-    <Converters:ThicknessToDoubleConverter x:Key="ThicknessToDoubleConverter" />
     <Converters:StringToVisibilityConverter x:Key="StringToVisibilityConverter" />
     <Converters:StringToVisibilityConverter x:Key="StringToOppositeVisibilityConverter" OppositeStringValue="True" />
