+++ src/MahApps.Metro/MahApps.Metro/Styles/Controls.Slider.xaml
@@ -1,15 +1,12 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:System="clr-namespace:System;assembly=mscorlib"
-                    xmlns:controls="clr-namespace:MahApps.Metro.Controls"
-                    xmlns:converters="clr-namespace:MahApps.Metro.Converters">
+                    xmlns:controls="clr-namespace:MahApps.Metro.Controls">
 
     <System:Double x:Key="HorizontalThumbHeight">16</System:Double>
     <System:Double x:Key="HorizontalThumbWidth">10</System:Double>
     <System:Double x:Key="VerticalThumbWidth">16</System:Double>
     <System:Double x:Key="VerticalThumbHeight">10</System:Double>
-
-    <converters:ThicknessToDoubleConverter x:Key="ThicknessToDouble" />
 
     <!--  HorizontalSliderThumb  -->
     <Style x:Key="HorizontalSliderThumb" TargetType="Thumb">
